// server/src/utils/enums/response-messages.enum.ts
export enum ResponseMessages {
  Success = 'Успешно',
  UserNotFound = 'Такого пользователя не существует',
  PostCreated = 'Запись создана',
  Registered = 'Регистраци прошла успешно',
  WrongPassword = 'Неверный пароль',
}